requirements.txt
streamlit
python-dotenv
langchain
langchain_openai
PyPDF2
chromadb
faiss-cpu
